Updates
Now, if the user uses SQLite, the plugin will inform them that they must use a MySQL database to use the plugin instead of attempting to run the MySQL queries. Message will display stating that the player needs to specify details to a MySQL server in the clockwork.cfg file.